LayoutTests/fast/frames/about-blank-frame-no-lazy-loading-expected.txt

Tests that it is possible to access frame.contentWindow synchronously on an about:blank iframe, even if marked as lazily loading.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframe.contentWindow.location.href is "about:blank"
PASS successfullyParsed is true

TEST COMPLETE
